[FEAT/#53] TerningDialog Column 삭제
