+++ core/src/main/res/drawable/ic_dialog_x_32.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="32dp"
+    android:height="32dp"
+    android:viewportWidth="32"
+    android:viewportHeight="32">
+  <group>
+    <clip-path
+        android:pathData="M0,0h32v32h-32z"/>
+    <path
+        android:pathData="M9.511,22.247C9.079,21.815 9.089,21.081 9.511,20.66L14.403,15.758L9.511,10.866C9.089,10.444 9.079,9.711 9.511,9.279C9.943,8.847 10.676,8.857 11.098,9.269L16,14.171L20.892,9.279C21.314,8.857 22.047,8.847 22.479,9.279C22.911,9.711 22.901,10.444 22.479,10.866L17.587,15.758L22.479,20.66C22.901,21.081 22.911,21.815 22.479,22.247C22.047,22.679 21.314,22.669 20.892,22.247L16,17.355L11.098,22.247C10.676,22.669 9.943,22.679 9.511,22.247Z"
+        android:fillColor="#BCBCBC"/>
+  </group>
+</vector>
