--- feature/src/main/java/com/terning/feature/dialog/TerningDialog.kt
@@ -1,73 +1,0 @@
-package com.terning.feature.dialog
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.statusBarsPadding
-import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Card
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
-import androidx.compose.ui.window.DialogProperties
-import com.terning.core.designsystem.theme.BackgroundColor
-
-@Composable
-fun TerningDialog(
-    onDismissRequest: () -> Unit,
-    properties: DialogProperties = DialogProperties(
-        usePlatformDefaultWidth = false,
-        decorFitsSystemWindows = true,
-        dismissOnBackPress = true,
-        dismissOnClickOutside = true,
-    ),
-    content: @Composable () -> Unit,
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(BackgroundColor),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Dialog(
-            onDismissRequest = { onDismissRequest() },
-            properties = properties,
-        ) {
-            content()
-        }
-    }
-}
-
-@Composable
-fun DialogContent() {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(200.dp)
-            .padding(16.dp),
-        shape = RoundedCornerShape(20.dp),
-    ) {
-        Text(
-            text = "This is a minimal dialog",
-            modifier = Modifier
-                .fillMaxSize()
-                .wrapContentSize(Alignment.Center),
-            textAlign = TextAlign.Center,
-        )
-    }
-}
-
-@Preview(showBackground = true, showSystemUi = true)
-@Composable
-fun TerningDialogPreview() {
-    TerningDialog(onDismissRequest = {}, content = { DialogContent() })
-}
