+++ feature/src/main/java/com/terning/feature/onboarding/signin/SignInRoute.kt
@@ -23,8 +23,8 @@ import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.flowWithLifecycle
 import androidx.navigation.NavHostController
-import com.terning.feature.dialog.DialogContent
-import com.terning.feature.dialog.TerningDialog
+import com.terning.core.designsystem.component.dialog.DialogContent
+import com.terning.core.designsystem.component.dialog.TerningDialog
 import com.terning.core.designsystem.theme.TerningPointTheme
 import com.terning.core.extension.toast
 import com.terning.feature.R
@@ -88,7 +88,11 @@ fun SignInScreen(
         openDialog.value -> {
             TerningDialog(
                 onDismissRequest = { openDialog.value = false },
-                content = { DialogContent() }
+                content = {
+                    DialogContent(
+                        onDismissRequest = { openDialog.value = false }
+                    )
+                }
             )
         }
     }
